add cleanup hours
